AST: Move SourceEntityWalker to IDE
