+++ include/swift/AST/ASTNode.h
@@ -18,7 +18,6 @@
 #define SWIFT_AST_AST_NODE_H
 
 #include "llvm/ADT/PointerUnion.h"
-#include "swift/AST/SourceEntityWalker.h"
 #include "swift/AST/TypeAlignments.h"
 
 namespace swift {
@@ -47,9 +46,6 @@ namespace swift {
 
     void walk(ASTWalker &Walker);
     void walk(ASTWalker &&walker) { walk(walker); }
-
-    void walk(SourceEntityWalker &Walker);
-    void walk(SourceEntityWalker &&walker) { walk(walker); }
 
     /// \brief get the underlying entity as a decl context if it is one,
     /// otherwise, return nullptr;
