+++ lib/IDE/CMakeLists.txt
@@ -12,6 +12,7 @@ add_swift_library(swiftIDE STATIC
   Utils.cpp
   IDETypeChecking.cpp
   APIDigesterData.cpp
+  SourceEntityWalker.cpp
   DEPENDS
     swift-syntax-generated-headers
   LINK_LIBRARIES
