+++ lib/Migrator/TypeOfMigratorPass.cpp
@@ -16,6 +16,7 @@
 #include "swift/AST/NameLookup.h"
 #include "swift/AST/ParameterList.h"
 #include "swift/AST/Types.h"
+#include "swift/IDE/SourceEntityWalker.h"
 #include "swift/Migrator/ASTMigratorPass.h"
 #include "swift/Parse/Lexer.h"
