+++ lib/Sema/InstrumenterSupport.h
@@ -17,6 +17,8 @@
 
 #include "TypeChecker.h"
 
+#include "swift/AST/ASTWalker.h"
+
 namespace swift {
 namespace instrumenter_support {
